Updated CRUD generation for protobuf
