+++ rice-box.go
@@ -40,23 +40,23 @@ func init() {
 	}
 	filea := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/delete_single.go.tmpl",
-		FileModTime: time.Unix(1527182442, 0),
-		Content:     string("\n// Delete deletes a {{.EntityName}} record from database and sets id to nil\nfunc (repo {{.EntityName}}Repository) Delete(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr  error\n\t\tstmt *sql.Stmt\n\t)\n\tid := *entity.ID\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tstmt, err = tx.Prepare(\"DELETE FROM {{.Table}} WHERE id = $1\")\n\t\tif err != nil {\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPreDelete(id, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\t_, err = stmt.Exec(id)\n\t\tif err == nil {\n\t\t\tentity.ID = nil\n\t\t} else {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPostDelete(id, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\t\n\treturn entity, nil\n}"),
+		FileModTime: time.Unix(1527558068, 0),
+		Content:     string("\n// Delete deletes a {{.EntityName}} record from database and sets id to nil\nfunc (repo {{.EntityName}}Repository) Delete(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr  error\n\t\tstmt *sql.Stmt\n\t)\n\tid := entity.ID\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tstmt, err = tx.Prepare(\"DELETE FROM {{.Table}} WHERE id = $1\")\n\t\tif err != nil {\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPreDelete(id, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\t_, err = stmt.Exec(id)\n\t\tif err == nil {\n\t\t\tentity.ID = {{pkeyPropertyEmptyVal .PrimaryKey}}\n\t\t} else {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPostDelete(id, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\t\n\treturn entity, nil\n}"),
 	}
 	fileb := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/get.go.tmpl",
-		FileModTime: time.Unix(1527506255, 0),
-		Content:     string("\n// Get returns a single {{.EntityName}} from database by primary key\nfunc (repo {{.EntityName}}Repository) Get(ctx context.Context, id {{pkeyPropertyType .PrimaryKey}}) ({{.EntityName}}, error) {\n\tvar (\n\t\trows   *sql.Rows\n\t\terr    error\n\t\tentity {{.EntityName}}\n\t)\n\t{{if .HasPreHook}}\n    if err = crudPreGet(ctx, id); err != nil {\n\t\treturn entity, err\n\t}\n    {{end}}\n\tselect {\n\tcase <- ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\trows, err = repo.db.Query(\"SELECT {{.SQLFields}} FROM {{.Table}} WHERE id = $1 ORDER BY .id ASC\", id)\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\tdefer rows.Close()\n\tif rows.Next() {\n\t\tselect {\n\t\tcase <- ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\t{{range .Before}}\n\t\t\t{{.}}{{end}}\n\t\t\terr = rows.Scan({{.StructFields}})\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t\t{{range .After}}\n\t\t\t{{.}}{{end}}\n\t\t} \n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPostGet(ctx, entity); err != nil {\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\n\treturn entity, nil\n}"),
+		FileModTime: time.Unix(1527538497, 0),
+		Content:     string("\n// Get returns a single {{.EntityName}} from database by primary key\nfunc (repo {{.EntityName}}Repository) Get(ctx context.Context, id {{pkeyPropertyType .PrimaryKey}}) ({{.EntityName}}, error) {\n\tvar (\n\t\trows   *sql.Rows\n\t\terr    error\n\t\tentity {{.EntityName}}\n\t)\n\t{{if .HasPreHook}}\n    if err = crudPreGet(ctx, id); err != nil {\n\t\treturn entity, err\n\t}\n    {{end}}\n\tselect {\n\tcase <- ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\trows, err = repo.db.Query(\"SELECT {{.SQLFields}} FROM {{.Table}} WHERE id = $1 ORDER BY .id ASC\", id)\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\tdefer rows.Close()\n\tif rows.Next() {\n\t\tselect {\n\t\tcase <- ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\tbreak\n\t\t}\n\n\t\t{{range .Before}}{{.}}{{end}}\n\n\t\terr = rows.Scan({{.StructFields}})\n\t\tif err != nil {\n\t\t\treturn entity, err\n\t\t}\n\t\t\n\t\t{{range .After}}{{.}}{{end}}\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPostGet(ctx, entity); err != nil {\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\n\treturn entity, nil\n}"),
 	}
 	filec := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/insert.go.tmpl",
-		FileModTime: time.Unix(1527506463, 0),
-		Content:     string("\n// Insert performs an SQL insert for {{.EntityName}} record and update instance with inserted id.\nfunc (repo {{.EntityName}}Repository) Insert(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\t{{- if pkeyIsAuto .PrimaryKey -}}\n\t\tid  {{pkeyPropertyType .PrimaryKey}}\n\t\t{{- end}}\n\t\terr  error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\t{{range .Before}}{{.}}{{end}}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\t{{if eq .PrimaryKey \"serial\" -}}\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.SQLFields}}) VALUES ({{.SQLPlaceholders}}) RETURNIentity, NG id\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\t{{else}}\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.SQLFields}}) VALUES ({{.SQLPlaceholders}})\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\t{{- end}}\n\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPreSave(\"INSERT\", entity, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t{{if eq .PrimaryKey \"serial\" -}}\n\terr = stmt.QueryRow({{.StructFields}}).Scan(&id)\n\tif err == nil {\n\t\tentity.ID = &id\n\t} else {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{else}}\n\t{{if eq .PrimaryKey \"uuid\" -}}\n\tidUUID, err := uuid.NewV4()\n\t\n\tif err == nil {\n\t\tid = idUUID.String()\n\t} else {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t*entity.ID = id\n\t{{- end}}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <- ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\tif err := crudPostSave(\"INSERT\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\n\tselect {\n\tcase <- ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, nil\n}"),
+		FileModTime: time.Unix(1527559583, 0),
+		Content:     string("\n// Insert performs an SQL insert for {{.EntityName}} record and update instance with inserted id.\nfunc (repo {{.EntityName}}Repository) Insert(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\t{{- if pkeyIsAuto .PrimaryKey -}}\n\t\tid  {{pkeyPropertyType .PrimaryKey}}\n\t\t{{- end}}\n\t\terr  error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\t{{range .Before}}{{.}}{{end}}\n\n\t{{if eq .PrimaryKey \"serial\" -}}\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.SQLFields}}) VALUES ({{.SQLPlaceholders}}) RETURNIentity, NG id\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\t{{else}}\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.SQLFields}}) VALUES ({{.SQLPlaceholders}})\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\t{{- end}}\n\n\t{{range .After}}{{.}}{{end}}\n\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif err = crudPreSave(\"INSERT\", entity, tx); err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn entity, err\n\t\t}\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t{{if eq .PrimaryKey \"serial\" -}}\n\terr = stmt.QueryRow({{.StructFields}}).Scan(&id)\n\tif err == nil {\n\t\tentity.ID = &id\n\t} else {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{else}}\n\t{{if eq .PrimaryKey \"uuid\" -}}\n\tidUUID, err := uuid.NewV4()\n\t\n\tif err == nil {\n\t\tid = idUUID.String()\n\t} else {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\tentity.ID = id\n\t{{- end}}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <- ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\tif err := crudPostSave(\"INSERT\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\n\tselect {\n\tcase <- ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, nil\n}"),
 	}
 	filed := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/list.go.tmpl",
-		FileModTime: time.Unix(1527506304, 0),
-		Content:     string("\n// List returns a slice containing {{.EntityName}} records\nfunc (repo {{.EntityName}}Repository) List(ctx context.Context, filters []ListFilter, offset, limit int) ([]{{.EntityName}}, error) {\n\tvar (\n\t\tlist     []{{.EntityName}}\n\t\tsegments []string\n\t\tvalues   []interface{}\n\t\terr      error\n\t\trows     *sql.Rows\n\t)\n\n\tquery := \"SELECT {{.SQLFields}} FROM {{.Table}}\"\n\t{{if .HasPreHook}}\n    if filters, err = crudPreList(ctx, filters); err != nil {\n\t\treturn nil, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn nil, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tfor i, filter := range filters {\n\t\tsegments = append(segments, filter.Field+\" \"+filter.Operation+\" $\"+strconv.Itoa(i+1))\n\t\tvalues = append(values, filter.Value)\n\t}\n\n\tif len(segments) != 0 {\n\t\tquery += \" WHERE \" + strings.Join(segments, \" AND \")\n\t}\n\n\tif limit > -1 {\n\t\tquery += \" LIMIT \"+strconv.Itoa(limit)\n\t}\n\n\tif offset > -1 {\n\t\tquery += \" OFFSET \"+strconv.Itoa(limit)\n\t}\n\n\tquery += \" ORDER BY id ASC\"\n\n\trows, err = repo.db.Query(query, values...)\n\tif err != nil {\n\t\treturn nil, err\n\t}\n\n\tdefer rows.Close()\n\tfor rows.Next() {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn nil, ctx.Err()\n\t\tdefault:\n\t\t\tbreak\n\t\t}\n\n\t\tvar entity {{.EntityName}}\n\t\terr = rows.Scan({{.StructFields}})\n\t\tif err != nil {\n\t\t\treturn nil, err\n\t\t}\n\n\t\tlist = append(list, entity)\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn nil, ctx.Err()\n\tdefault:\n\t\tif list, err = crudPostList(ctx, list); err != nil {\n\t\t\treturn nil, err\n\t\t}\n\t}\n\t{{end}}\n\treturn list, nil\n}"),
+		FileModTime: time.Unix(1527556987, 0),
+		Content:     string("\n// List returns a slice containing {{.EntityName}} records\nfunc (repo {{.EntityName}}Repository) List(ctx context.Context, filters []ListFilter, offset, limit int) ([]{{.EntityName}}, error) {\n\tvar (\n\t\tlist     []{{.EntityName}}\n\t\tsegments []string\n\t\tvalues   []interface{}\n\t\terr      error\n\t\trows     *sql.Rows\n\t)\n\n\tquery := \"SELECT {{.SQLFields}} FROM {{.Table}}\"\n\t{{if .HasPreHook}}\n    if filters, err = crudPreList(ctx, filters); err != nil {\n\t\treturn nil, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn nil, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tfor i, filter := range filters {\n\t\tsegments = append(segments, filter.Field+\" \"+filter.Operation+\" $\"+strconv.Itoa(i+1))\n\t\tvalues = append(values, filter.Value)\n\t}\n\n\tif len(segments) != 0 {\n\t\tquery += \" WHERE \" + strings.Join(segments, \" AND \")\n\t}\n\n\tif limit > -1 {\n\t\tquery += \" LIMIT \"+strconv.Itoa(limit)\n\t}\n\n\tif offset > -1 {\n\t\tquery += \" OFFSET \"+strconv.Itoa(limit)\n\t}\n\n\tquery += \" ORDER BY id ASC\"\n\n\trows, err = repo.db.Query(query, values...)\n\tif err != nil {\n\t\treturn nil, err\n\t}\n\n\tdefer rows.Close()\n\tfor rows.Next() {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn nil, ctx.Err()\n\t\tdefault:\n\t\t\tbreak\n\t\t}\n\n\t\tvar entity {{.EntityName}}\n\t\t{{range .Before}}{{.}}{{end}}\n\n\t\terr = rows.Scan({{.StructFields}})\n\t\tif err != nil {\n\t\t\treturn nil, err\n\t\t}\n\t\t\n\t\t{{range .After}}{{.}}{{end}}\n\n\t\tlist = append(list, entity)\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn nil, ctx.Err()\n\tdefault:\n\t\tif list, err = crudPostList(ctx, list); err != nil {\n\t\t\treturn nil, err\n\t\t}\n\t}\n\t{{end}}\n\treturn list, nil\n}"),
 	}
 	filee := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/loadrelated.go.tmpl",
@@ -65,27 +65,27 @@ func init() {
 	}
 	filef := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/merge.go.tmpl",
-		FileModTime: time.Unix(1527182527, 0),
-		Content:     string("\n// Merge performs an SQL merge for {{.EntityName}} record.\nfunc (repo {{.EntityName}}Repository) Merge(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\n\tif entity.ID == nil {\n\t\treturn {{.EntityName}}Repo.Insert(ctx, entity, tx, autocommit)\n\t}\n\n\t{{range .Before}}{{.}}{{end}}\n\n\tstmt, err = tx.Prepare(`INSERT INTO {{.Table}} ({{.SQLFieldsInsert}}) VALUES ({{.SQLPlaceholders}}) \n\tON CONFLICT (id) DO UPDATE SET {{.SQLFieldsUpdate}}`)\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t{{if .HasPreHook}}\n    if err = crudPreSave(\"MERGE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\tif err = crudPostSave(\"MERGE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, err\n}"),
+		FileModTime: time.Unix(1527560386, 0),
+		Content:     string("\n// Merge performs an SQL merge for {{.EntityName}} record.\nfunc (repo {{.EntityName}}Repository) Merge(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\n\tif entity.ID == {{pkeyPropertyEmptyVal .PrimaryKey}} {\n\t\treturn {{.EntityName}}Repo.Insert(ctx, entity, tx, autocommit)\n\t}\n\n\t{{range .Before}}{{.}}{{end}}\n\n\tstmt, err = tx.Prepare(`INSERT INTO {{.Table}} ({{.SQLFieldsInsert}}) VALUES ({{.SQLPlaceholders}}) \n\tON CONFLICT (id) DO UPDATE SET {{.SQLFieldsUpdate}}`)\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t{{if .HasPreHook}}\n    if err = crudPreSave(\"MERGE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\n\t{{range .After}}{{.}}{{end}}\n\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\tif err = crudPostSave(\"MERGE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, err\n}"),
 	}
 	fileg := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/save.go.tmpl",
-		FileModTime: time.Unix(1527182266, 0),
-		Content:     string("\n// Save either inserts or updates a {{.EntityName}} record based on whether or not id is nil\nfunc (repo {{.EntityName}}Repository) Save(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\t{{if pkeyIsAuto .PrimaryKey -}}\n\tif entity.ID == nil {\n\t\treturn {{.EntityName}}Repo.Insert(ctx, entity, tx, autocommit)\n\t}\n\treturn {{.EntityName}}Repo.Update(ctx, entity, tx, autocommit)\n\t{{- else -}}\n\tif entity.ID == nil {\n\t\treturn entity, errors.New(\"primary key cannot be nil\")\n\t}\n\treturn {{.EntityName}}Repo.Merge(ctx, entity, tx, autocommit)\n\t{{end -}}\n}"),
+		FileModTime: time.Unix(1527558144, 0),
+		Content:     string("\n// Save either inserts or updates a {{.EntityName}} record based on whether or not id is nil\nfunc (repo {{.EntityName}}Repository) Save(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\t{{if pkeyIsAuto .PrimaryKey -}}\n\tif entity.ID == {{pkeyPropertyEmptyVal .PrimaryKey}} {\n\t\treturn {{.EntityName}}Repo.Insert(ctx, entity, tx, autocommit)\n\t}\n\treturn {{.EntityName}}Repo.Update(ctx, entity, tx, autocommit)\n\t{{- else -}}\n\tif entity.ID == {{pkeyPropertyEmptyVal .PrimaryKey}} {\n\t\treturn entity, errors.New(\"primary key cannot be nil\")\n\t}\n\treturn {{.EntityName}}Repo.Merge(ctx, entity, tx, autocommit)\n\t{{end -}}\n}"),
 	}
 	fileh := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/saverelated.go.tmpl",
-		FileModTime: time.Unix(1526938054, 0),
-		Content:     string("// SaveRelated{{.PropertyName}} is a helper function to save related {{.PropertyName}} entities\nfunc SaveRelated{{.PropertyName}}(ctx context.Context, idthis  {{pkeyPropertyType .PrimaryKey}}, relatedids {{.PropertyType}}) error {\n\tvar (\n\t\tplaceholder string\n\t\tvalues  []interface{}\n\t\tindices map[{{pkeyPropertyType .PrimaryKey}}]{{trimPrefix .PropertyType \"[]\"}}\n\t\tidthis  {{pkeyPropertyType .PrimaryKey}}\n\t\tidthat  {{trimPrefix .PropertyType \"[]\"}}\n\t\titems   []{{.EntityName}}\n\t\tstmt    *sql.Stmt\n\t)\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t\n\tstmt, err = tx.Prepare(\"DELETE FROM {{.Table}} WHERE {{.ThisID}} = $1\")\n\n\tif err != nil {\n\t\treturn err\n\t}\n\n\t_, err = stmt.Exec(*entity.ID)\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn err\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.ThisID}}, {{.ThatID}}) VALUES ($1, $2)\")\n\tif err != nil {\n\t\treturn err\n\t}\n\n\tfor _, relatedID := range *entity.{{.PropertyName}} {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\ttx.Rollback()\n\t\t\treturn ctx.Err()\n\t\tdefault:\n\t\t\tbreak\n\t\t}\n\n\t\t_, err = stmt.Exec(*entity.ID, relatedID)\n\t\tif err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn err\n\t\t}\n\t}\n}"),
+		FileModTime: time.Unix(1527557896, 0),
+		Content:     string("// SaveRelated{{.PropertyName}} is a helper function to save related {{.PropertyName}} entities\nfunc SaveRelated{{.PropertyName}}(ctx context.Context, idthis  {{pkeyPropertyType .PrimaryKey}}, relatedids {{.PropertyType}}) error {\n\tvar (\n\t\tplaceholder string\n\t\tvalues  []interface{}\n\t\tindices map[{{pkeyPropertyType .PrimaryKey}}]{{trimPrefix .PropertyType \"[]\"}}\n\t\tidthis  {{pkeyPropertyType .PrimaryKey}}\n\t\tidthat  {{trimPrefix .PropertyType \"[]\"}}\n\t\titems   []{{.EntityName}}\n\t\tstmt    *sql.Stmt\n\t)\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t\n\tstmt, err = tx.Prepare(\"DELETE FROM {{.Table}} WHERE {{.ThisID}} = $1\")\n\n\tif err != nil {\n\t\treturn err\n\t}\n\n\t_, err = stmt.Exec(entity.ID)\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn err\n\t}\n\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tstmt, err = tx.Prepare(\"INSERT INTO {{.Table}} ({{.ThisID}}, {{.ThatID}}) VALUES ($1, $2)\")\n\tif err != nil {\n\t\treturn err\n\t}\n\n\tfor _, relatedID := range *entity.{{.PropertyName}} {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\ttx.Rollback()\n\t\t\treturn ctx.Err()\n\t\tdefault:\n\t\t\tbreak\n\t\t}\n\n\t\t_, err = stmt.Exec(entity.ID, relatedID)\n\t\tif err != nil {\n\t\t\ttx.Rollback()\n\t\t\treturn err\n\t\t}\n\t}\n}"),
 	}
 	filei := &embedded.EmbeddedFile{
 		Filename:    "crud/partials/update.go.tmpl",
-		FileModTime: time.Unix(1527506677, 0),
-		Content:     string("\n// Update Will execute an SQLUpdate Statement for {{.EntityName}} in the database. Prefer using Save instead of Update directly.\nfunc (repo {{.EntityName}}Repository) Update(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\t{{range .Before}}{{.}}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tstmt, err = tx.Prepare(\"UPDATE {{.Table}} SET {{.SQLFields}} WHERE id = $1\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n    if err = crudPreSave(\"UPDATE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n        return entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tif err = crudPostSave(\"UPDATE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, err\n}"),
+		FileModTime: time.Unix(1527559977, 0),
+		Content:     string("\n// Update Will execute an SQLUpdate Statement for {{.EntityName}} in the database. Prefer using Save instead of Update directly.\nfunc (repo {{.EntityName}}Repository) Update(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {\n\tvar (\n\t\terr error\n\t\tstmt *sql.Stmt\n\t)\n\n\tif tx == nil {\n\t\tselect {\n\t\tcase <-ctx.Done():\n\t\t\treturn entity, ctx.Err()\n\t\tdefault:\n\t\t\ttx, err = repo.db.Begin()\n\t\t\tif err != nil {\n\t\t\t\treturn entity, err\n\t\t\t}\n\t\t}\n\t}\n\t\n\t{{range .Before}}{{.}}{{end}}\n\n\tstmt, err = tx.Prepare(\"UPDATE {{.Table}} SET {{.SQLFields}} WHERE id = $1\")\n\tif err != nil {\n\t\treturn entity, err\n\t}\n\n\t{{range .After}}{{.}}{{end}}\n\n\t{{if .HasPreHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n    if err = crudPreSave(\"UPDATE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n        return entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\t_, err = stmt.Exec({{.StructFields}})\n\tif err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{if .HasPostHook}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tbreak\n\t}\n\n\tif err = crudPostSave(\"UPDATE\", entity, tx); err != nil {\n\t\ttx.Rollback()\n\t\treturn entity, err\n\t}\n\t{{end}}\n\tselect {\n\tcase <-ctx.Done():\n\t\ttx.Rollback()\n\t\treturn entity, ctx.Err()\n\tdefault:\n\t\tif autocommit {\n\t\t\terr = tx.Commit()\n\t\t}\n\t}\n\n\treturn entity, err\n}"),
 	}
 	filej := &embedded.EmbeddedFile{
 		Filename:    "crud/protobuf.proto.tmpl",
-		FileModTime: time.Unix(1527537690, 0),
+		FileModTime: time.Unix(1527556750, 0),
 		Content:     string("syntax = \"proto3\";\n\npackage models;\n\n{{range .Imports -}}\n{{.}}\n{{- end}}\n\n{{ range .Entities -}}\n// {{.Name}} {{.Description}}\nmessage {{.Name}} { {{ range .Fields }}\n\t{{.Type}} {{.Name}} = {{.Index}};\n{{- end}}\n}\n\n{{end}}"),
 	}
 	filek := &embedded.EmbeddedFile{
