+++ templates/crud/partials/save.go.tmpl
@@ -2,12 +2,12 @@
 // Save either inserts or updates a {{.EntityName}} record based on whether or not id is nil
 func (repo {{.EntityName}}Repository) Save(ctx context.Context, entity {{.EntityName}}, tx *sql.Tx, autocommit bool) ({{.EntityName}}, error) {
 	{{if pkeyIsAuto .PrimaryKey -}}
-	if entity.ID == nil {
+	if entity.ID == {{pkeyPropertyEmptyVal .PrimaryKey}} {
 		return {{.EntityName}}Repo.Insert(ctx, entity, tx, autocommit)
 	}
 	return {{.EntityName}}Repo.Update(ctx, entity, tx, autocommit)
 	{{- else -}}
-	if entity.ID == nil {
+	if entity.ID == {{pkeyPropertyEmptyVal .PrimaryKey}} {
 		return entity, errors.New("primary key cannot be nil")
 	}
 	return {{.EntityName}}Repo.Merge(ctx, entity, tx, autocommit)
